small fix to avoid compiler warnings
git-svn-id: svn://tug.org/texlive/trunk/Build/source@13965 c570f23f-e606-0410-a88d-b1316a301751

+++ libs/teckit/source/Compiler.h
@@ -230,7 +230,7 @@ protected:
 		vector<UInt32>		byteClassLines;
 		vector<UInt32>		uniClassLines;
 
-		long				passType;
+		UInt32				passType;
 		UInt32				uniDefault;
 		UInt8				byteDefault;
 		bool				supplementaryChars;
